requirements.txt
fastapi
uvicorn[standard]
openai
anthropic
httpx
python-dotenv
pytest
